Sale operation

+++ src/main/java/com/khem/appspring/springphoneshop/dto/PriceDTO.java
@@ -1,8 +1,10 @@
 package com.khem.appspring.springphoneshop.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class PriceDTO {
-	private double salePrice;
+	private BigDecimal salePrice;
 }
